src/agents/requirements.txt
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
supabase>=1.0.0
httpx>=0.24.0
typing-extensions>=4.0.0
tenacity>=8.0.0